refactor: questionnaire path
