+++ app/questionnaire/[id]/components/(question-types)/(multiple-choice-question)/index.tsx
@@ -1,6 +1,6 @@
-import { MultipleChoiceSubType } from "@/app/questionnaire/[id]/question/types/multiple-choice-types";
+import { MultipleChoiceSubType } from "@/app/questionnaire/[id]/types/multiple-choice-types";
 import DCheckbox from "./DCheckbox";
-import { ComponentRegistry } from "@/app/questionnaire/[id]/question/types/component-types";
+import { ComponentRegistry } from "@/app/questionnaire/[id]/types/component-types";
 import { QuestionComponent } from "../../QuestionComponent";
 
 const MULTI_CHOICE_COMPONENTS: ComponentRegistry<MultipleChoiceSubType> = {
