+++ app/questionnaire/[id]/components/(question-types)/(text-question)/index.tsx
@@ -1,8 +1,8 @@
-import { TextSubType } from "@/app/questionnaire/[id]/question/types/text-types";
+import { TextSubType } from "@/app/questionnaire/[id]/types/text-types";
 import ShortText from "./ShortText";
 import LongText from "./LongText";
 import Number from "./Number";
-import { ComponentRegistry } from "@/app/questionnaire/[id]/question/types/component-types";
+import { ComponentRegistry } from "@/app/questionnaire/[id]/types/component-types";
 import { QuestionComponent } from "../../QuestionComponent";
 
 const TEXT_COMPONENTS: ComponentRegistry<TextSubType> = {
