+++ app/questionnaire/[id]/components/QuestionCard.tsx
@@ -10,11 +10,11 @@ import React from "react";
 import SingleChoiceQuestion from "./(question-types)/(single-choice-question)";
 import MultipleChoiceQuestion from "./(question-types)/(multiple-choice-question)";
 import TextQuestion from "./(question-types)/(text-question)";
-import { QuestionType } from "@/app/questionnaire/[id]/question/types/question-types";
+import { QuestionType } from "@/app/questionnaire/[id]/types/question-types";
 import InvalidComponent from "./InvalidComponent";
 import { useQuestionContext } from "../context/question-context";
 import { useCallback } from "react";
-import { useConditionalLogic } from "@/app/questionnaire/[id]/question/hooks/useConditionalLogic";
+import { useConditionalLogic } from "@/app/questionnaire/[id]/hooks/useConditionalLogic";
 import SubmitAlert from "./SubmitAlert";
 import { useRouter } from "next/navigation";
 import { t } from "@/app/locales/translation";
